Simplify extract function creation
Having two named functions for each extract variant is nice, but it's
silly how verbose it was to write them, and the previous solution with
partial application wasn't great. It just pushed the duplication up one
level.

Here I added a macro for automating the creation of both variants of an
extract function, and also made it so that only the path transformation
part of the code needs to be written.